refactor: move `CompressionAlgorithm` in a dedicated module
