+++ mithril-common/src/entities/mod.rs
@@ -13,6 +13,7 @@ mod cardano_transactions_snapshot;
 mod certificate;
 mod certificate_metadata;
 mod certificate_pending;
+mod compression_algorithm;
 mod epoch;
 mod http_server_error;
 mod mithril_stake_distribution;
@@ -40,6 +41,7 @@ pub use cardano_transactions_snapshot::CardanoTransactionsSnapshot;
 pub use certificate::{Certificate, CertificateSignature};
 pub use certificate_metadata::{CertificateMetadata, StakeDistributionParty};
 pub use certificate_pending::CertificatePending;
+pub use compression_algorithm::*;
 pub use epoch::{Epoch, EpochError};
 pub use http_server_error::{ClientError, ServerError};
 pub use mithril_stake_distribution::MithrilStakeDistribution;
@@ -51,6 +53,6 @@ pub use signed_entity_type::*;
 pub use signer::{Signer, SignerWithStake};
 pub use single_signatures::*;
 pub use slot_number::SlotNumber;
-pub use snapshot::{CompressionAlgorithm, Snapshot};
+pub use snapshot::Snapshot;
 pub use time_point::*;
 pub use type_alias::*;
